Fixed File Paths-- Organized Pages and Assets
