+++ my-app/src/Pages/LoginSignup.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react'
 import './LoginSignup.css'
 
-import user_icon from "../Assets/person.png"
-import email_icon from "../Assets/email.png"
-import password_icon from "../Assets/password.png"
+import user_icon from "../Components/Assets/person.png"
+import email_icon from "../Components/Assets/email.png"
+import password_icon from "../Components/Assets/password.png"
 
 const LoginSignup = () => {
